+++ src/lib/domain/nutritions.dart
@@ -1,5 +1,5 @@
 class Nutritions {
-  Nutritions({required int kJoule, double? carbohydrates, double? sugar, double? fat, double? saturatedFat, double? protein, double? salt})
+  Nutritions({required double kJoule, double? carbohydrates, double? sugar, double? fat, double? saturatedFat, double? protein, double? salt})
     : _kJoule = kJoule,
       _carbohydrates = carbohydrates,
       _sugar = sugar,
@@ -8,7 +8,7 @@ class Nutritions {
       _protein = protein,
       _salt = salt;
 
-  final int _kJoule;
+  final double _kJoule;
   final double? _carbohydrates;
   final double? _sugar;
   final double? _fat;
@@ -16,7 +16,7 @@ class Nutritions {
   final double? _protein;
   final double? _salt;
 
-  int get kJoule => _kJoule;
+  double get kJoule => _kJoule;
   double? get carbohydrates => _carbohydrates;
   double? get sugar => _sugar;
   double? get fat => _fat;
